+++ api/src/opentrons/protocol_engine/commands/flex_stacker/retrieve.py
@@ -14,7 +14,7 @@ from ..command import (
     SuccessData,
     DefinedErrorData,
 )
-from ..movement_common import FlexStackerStallOrCollisionError
+from ..flex_stacker.common import FlexStackerStallOrCollisionError
 from ...errors import (
     ErrorOccurrence,
     CannotPerformModuleAction,
