+++ api/src/opentrons/protocol_engine/commands/movement_common.py
@@ -115,22 +115,6 @@ class StallOrCollisionError(ErrorOccurrence):
     detail: str = ErrorCodes.STALL_OR_COLLISION_DETECTED.value.detail
 
 
-class FlexStackerStallOrCollisionError(ErrorOccurrence):
-    """Returned when the motor driver detects a stall."""
-
-    All axes are stopped at the point where the error was encountered.
-
-    The next thing to move the machine must account for the robot not having a valid estimate
-    of its position. It should be a `home` or `unsafe/updatePositionEstimators`.
-    """
-
-    isDefined: bool = True
-    errorType: Literal["flexStackerStallOrCollision"] = "flexStackerStallOrCollision"
-
-    errorCode: str = ErrorCodes.STALL_OR_COLLISION_DETECTED.value.code
-    detail: str = ErrorCodes.STALL_OR_COLLISION_DETECTED.value.detail
-
-
 class DestinationPositionResult(BaseModel):
     """Mixin for command results that move a pipette."""
